# api/openapi.yaml
openapi: 3.1.0
info:
  title: doenerstag
  version: "1"
  summary: Coordinate food orders for a group.
  description: |
    doenerstag coordinates food orders for a group. It does not take payments
    and does not place orders with restaurants.

    This document describes every endpoint the server registers. A test walks
    the routes of the production router and fails when one of them is missing
    from here, and fails the other way too when this document describes a path
    nothing serves — so the two cannot drift apart.

    Reading is mostly open. An unauthenticated visitor sees restaurants, menus,
    the list of orders and each order's item count, but never who ordered what:
    the split is enforced in the handlers and is described at each endpoint it
    affects.

    Collections are wrapped in an object with one plural key rather than being
    returned as a bare array — `{"restaurants": [...]}` — which leaves room to
    add a field beside the list without changing the type of the response.
  license:
    name: BSD-3-Clause
    identifier: BSD-3-Clause
servers:
  - url: /api/v1
    description: This installation.
...
security:
  - session: []
  - token: []
paths:
...
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: Every account. Signed in for the names, administrator for the rest.
      description: |
        Two shapes, chosen by who is asking. An administrator gets `AdminUser`,
        with the e-mail address and the login times. Anybody else who is signed
        in gets `User` -- id, user name, display name -- which is what
        `GET /users/{id}` already serves to anyone at all.

        The second tier exists so that an order can name accounts rather than
        typed names: both the person collecting the money and the person
        fetching the food are chosen from this list. What it adds over the
        per-user endpoint is enumeration, and that is why it needs a session.
      responses:
        "200":
          description: All accounts, in the shape the caller is entitled to.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      oneOf:
                        - $ref: "#/components/schemas/AdminUser"
                        - $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      operationId: getUser
      summary: A public profile.
      description: |
        Id, user name and display name for anybody; the e-mail address only to
        the account itself and to the administrator.
      security: []
      responses:
        "200":
          description: The profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [users]
      operationId: patchUser
      summary: Change an account. Owner or administrator.
      description: |
        Only the fields present in the body change. Changing your own password
        requires the current one — a borrowed unlocked browser should not be
        enough to lock the owner out — while the administrator resetting
        somebody else's does not, because that is the recovery path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserRequest"
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [users]
      operationId: deleteUser
      summary: Delete an account. Owner or administrator.
      description: |
        Immediate and irreversible, and it takes the account's items in active
        orders with it. Items in expired orders and orders the account created
        are reassigned to the deleted-user placeholder, so history stays
        readable. Ask `/deletion-impact` first.
      responses:
        "204":
          description: The account is gone.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/deletion-impact:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      operationId: getDeletionImpact
      summary: What deleting this account would cost.
      description: |
        The numbers the confirmation dialog names, so that a person confirms
        something specific rather than something general.
      responses:
        "200":
          description: The impact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletionImpact"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/tokens:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      operationId: listTokens
      summary: The account's API tokens.
      description: The value is never returned; only its prefix, for recognition.
      responses:
        "200":
          description: The tokens.
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/Token"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [users]
      operationId: createToken
      summary: Issue an API token.
      description: |
        The response carries the token value, and it is the only time it ever
        does: what is stored is a hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: The new token, with its value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedToken"
        "400":
          $ref: "#/components/responses/Validation"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/{id}/tokens/{tid}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - name: tid
        in: path
        required: true
        description: The token's id.
        schema: { type: string, format: uuid }
    delete:
      tags: [users]
      operationId: revokeToken
      summary: Revoke an API token.
      responses:
        "204":
          description: The token is revoked.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: doener_session
      description: |
        A JWT backed by a server-side session row. State-changing requests must
        also send the CSRF token from the doener_csrf cookie in the
        X-CSRF-Token header.
    token:
      type: http
      scheme: bearer
      description: |
        An API token for third-party and scripted access. Not subject to the
        CSRF check, because it carries no ambient browser credentials.
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: The account's id.
      schema: { type: string, format: uuid }
...
  responses:
...
    Validation:
      description: The request was refused. Errors 1000 to 1999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required, or the session has ended. Errors 2000 to 2999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The caller is authenticated but not permitted. Errors 3000 to 3999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such thing. Error 4000.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current state. Errors 4001 to 4999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      description: |
        Every failure carries this shape. The code is an internal, stable
        number the frontend keys its translations on; the message is
        developer-facing English and is never the primary way a user learns
        what went wrong.

        The request_id matches the X-Request-Id response header and the log
        line, which is what makes a report of "it went wrong just now"
        traceable without the response ever carrying SQL text or a stack trace.
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: integer
              description: The internal error number. See docs/04_api.md.
            message:
              type: string
            field:
              type: [string, "null"]
              description: The offending request field, for validation errors.
            request_id:
              type: string
    User:
      type: object
      required: [id, name, display_name, is_admin]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        display_name: { type: string }
        is_admin: { type: boolean }
        email:
          type: string
          description: Only present for the account itself and the administrator.
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [created_at]
          properties:
            created_at: { type: string, format: date-time }
            last_login_at: { type: string, format: date-time }
...
    PatchUserRequest:
      type: object
      description: Only the fields present are changed.
      properties:
        name: { type: string }
        display_name: { type: string }
        email: { type: string }
        password: { type: string }
        current_password:
          type: string
          description: |
            Required when changing your own password. The administrator
            resetting somebody else's does not supply it, because they do not
            have it.
    DeletionImpact:
      type: object
      required: [active_orders, active_items, expired_items, created_active_orders]
      properties:
        active_orders:
          type: integer
          description: Active orders this account has items in.
        active_items:
          type: integer
          description: Items that would be deleted.
        expired_items:
          type: integer
          description: Items that would be reassigned to the deleted user.
        created_active_orders:
          type: integer
          description: Active orders this account opened, which would be reassigned.
    Token:
      type: object
      required: [id, name, prefix, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix:
          type: string
          description: The first characters of the value, for recognition.
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
    CreatedToken:
      allOf:
        - $ref: "#/components/schemas/Token"
        - type: object
          required: [token]
          properties:
            token:
              type: string
              description: |
                The value, returned exactly once. What is stored is a hash, so
                this response is the only chance to copy it.
    CreateTokenRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 64 }
        expires_at:
          type: string
          description: RFC 3339, in the future. Empty means the token never expires.
...